% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\docType{data}
\name{data_list}
\alias{data_list}
\title{A list of longitudinal/clustered datasets}
\format{
A list containing 19 individual clustered datasets. Each dataset has a common outcome and clustering variable defined below:
\describe{
  \item{target}{The outcome variable of each dataset of the list}
  \item{cluster_id}{the clustering variable within each dataset of the list} 
}
}
\usage{
data_list
}
\description{
A list of longitudinal/clustered datasets
}
\keyword{datasets}
