% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{List of Longitduinal datasets}
\format{

}
\usage{
datasets
}
\description{
Datasets are names "datX" where "X" is a number. Currently there are 19 individual longitudinal datasets and a "meta_data" file summarizing them
}
\keyword{datasets}
