% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter datasets based on criteria in the metadata file}
\usage{
filter_data(..., subset = T)
}
\arguments{
\item{...}{dplyr `filter()' syntax for filtering data. Should be based on columns in the meta_data file}

\item{subset}{Logical (T or F). If T, then function returns list of dataframes matching filter criteria. If F, then function returns a vector of datasets that match filter criteria}
}
\value{
Dependent on 'subset' - A list of datasets matching filter criteria or a vector of dataset names matching filter criteria
}
\description{
Gives list of datasets meeting criteria specified in the function.
}
\examples{
subset_data_list= filter_data(n_features>10, n_obs>2500, subset=T) #returns a list of datasets with >10 features and >2500 observations
subset_data_names = filter_data(n_features>10, n_obs>2500, subset=T) #returns a vector of dataset names matching the filter criteria
}
