% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Dataset Extraction}
\usage{
get_data(x)
}
\arguments{
\item{x}{The dataset/filename as described in 'data_meta.csv" or "data_info.csv"}
}
\value{
The datasset asked for
}
\description{
Extract specific dataset by name.
}
\examples{
data_example <- get_data("dat1");
}
