% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_datasets.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Metadata file creation}
\usage{
get_meta_data()
}
\value{
A new datafile saved in the "data/" folder containing up-to-date meta data on all data sets
}
\description{
Pull all datasets from "data/" file, summarize them, and write them to new file "data_meta.csv"
}
\examples{
get_meta_data() 
}
