% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta_data.R
\name{plot_meta_data}
\alias{plot_meta_data}
\title{Provides plot of meta data for all datasets in the 'data_list'}
\usage{
plot_meta_data(allplots = T, df = OpenClustered::meta_data)
}
\arguments{
\item{allplots}{Logical (T or F); If T, returns a 4-panel grid of each plot. If F, returns a list of each of the 4 plots as an individual element}

\item{df}{the meta data file one returned from get_data("meta_data.csv");}
}
\value{
4-panel figure of metadata summary
}
\description{
Creates plots summarizing the meta data (n obervations, n features, n clusters, and imbalance) of the 19 datasets in the 'data_list'
}
\examples{
plot_meta_data(allplots=T)
}
