% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_datasets_smry.R
\name{vis_datasets_smry}
\alias{vis_datasets_smry}
\title{Visualize summary characteristics of all available datasets}
\usage{
vis_datasets_smry(df)
}
\arguments{
\item{df}{the meta data file one returned from get_data("meta_data.csv");}
}
\value{
4-panel figure of metadata summary
}
\description{
Creates a 4-panel figure of summary statistics of the dataset meta data, summarizes the total number of observations, number of features, number of clusters and imbalance.
}
\examples{
meta_data = get_data("meta_data")
vis_datasets_smry(meta_data)
}
