% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter datasets based on criteria in the metadata file}
\usage{
filter_data(...)
}
\arguments{
\item{...}{dplyr `filter()' syntax for filtering data. Should be based on columns in the meta_data file}
}
\value{
a vector of dataset names meeting the filtered criteria
}
\description{
Gives list of datasets meeting criteria specified in the function.
}
\examples{
subset_data = filter_data(n_features>10, n_obs>2500)
}
